// src/app/models/feedback.ts
export class Feedback {
    Timestamp: Date;
    Company: string;
    User: string;
    EnergyLayer: string;
    DataSet: string;
    Description: string;
    Clicklocation: string;
    ShapeName: string;
    NodeSqlId: string;
    MSID: string;
    ObjectId: string;
    Message: string;
    Host: string;
    IP: string;
    UserId: string;
    EnergyLayerGuid:string;
}